deleted unused files, tournamentsview working with JS requests, deleted unnecessary components, added packages in Dockerfile
